requirements.txt
flask
python-dotenv
requests
scikit-learn
numpy
pandas
textblob
nltk
vaderSentiment
matplotlib
seaborn
transformers
torch
